feat(#61): add group creation link and validations (#89)

+++ src/static/js/organizations.js
@@ -1,6 +1,10 @@
 (async () => {
   const deleteOrganization = async event => {
     event.preventDefault();
+    // NOTE: I know alerts are annoying, but we're out of time!
+    if (!confirm('Are you sure? This can not be undone.')) {
+      return;
+    }
     const orgId = event.target.getAttribute('data-id');
     const res = await fetch(`/organizations/${orgId}`, {
       method: 'DELETE',
